core-image-minimal-initramfs-alix3d3: add initramfs image
Add a specific initramfs image for ALIX3D3 board. The initramfs has
initramfs-framework-base and udev-module.

ALIX3D3 image is formatted with two partitions. The first has boot, and
the second partition has rootfs. In this case,
core-image-minimal-initramfs will not work as it depends on live init
scripts, which in turn looks for rootfs.img file and mounts it into
/dev/ram0