classes, alix3d3.conf: Add IMAGE_CMD_alix-hddimg command
Add a bbclass for creating hdd images with 2 partitions. The first
partition has syslinux, initrd, and kernel, the second partition has
rootfs. The produced image can be directly dd'ed into CF card and run.

Since /dev/sda2 is mounted as root, the changes to the system is
permanent. It is useful for daemon configuration (such as aprx).